requirements.txt
flasgger==0.9.5
Flask==1.1.2
Flask-Cors==3.0.8
python-decouple==3.3
pytest==6.0.1